requirements.txt
requests==2.31.0
redis==5.0.3
python-dotenv==1.0.1
urllib3==2.2.1
langchain==0.2.3
scikit-learn==1.5.0
langchain-community==0.2.4
openpyxl
pandas